+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* Upgrade of Integration Test Helper (0.0.4)
+* Upgrade to subclass of IntegrationTestHelperBinding
+* Added super() call for BaseIntegrationTest(binding)
+
 ## 0.0.1
 
 * TODO: Describe initial release.